pyproject.toml
[project]
name = "custa"
version = "0.1.1"
description = "Custom HTML generator from simple markup syntax"
readme = "README.md"
license = {text = "MIT"}
authors = [{ name = "ProPandaMen", email = "propandamen@gmail.com" }]
requires-python = ">=3.8"
dependencies = [
  "PyYAML>=6.0",
  "typer>=0.15",
]

[project.urls]
"Homepage" = "https://github.com/ProPandaMen/custa"

[project.scripts]
custa = "custa.cli:app"

[tool.setuptools]
packages = ["custa"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
